% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getidfun.R
\name{get_id}
\alias{get_id}
\title{Obtains recipe ID number with the input of ingredients}
\usage{
get_id(ingredients, nRecipes = 1)
}
\arguments{
\item{ingredients}{A character vector of individual ingredients;
no limit on length}

\item{nRecipes}{The number of recipes to retrieve. A whole number.
Default = 1}
}
\value{
Gives an ID number/s of recipes that use the listed ingredients
}
\description{
Obtains recipe ID number with the input of ingredients
}
\examples{
FoodYouHave <- c("chickpeas", "tomato", "spinach")
ids <- get_id(FoodYouHave, 3)

print(ids)

}
