% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infofun.R
\name{get_info}
\alias{get_info}
\title{Gives full recipe information with input of ingredients and dietary restriction}
\usage{
get_info(ingredients, diet = "")
}
\arguments{
\item{ingredients}{A character vector of individual ingredients;
no limit on length}

\item{diet}{A character of dietary restrictions. Can be 'vegetarian', 'vegan',
'glutenFree', or 'dairyFree'. By default it is empty.}
}
\value{
Gives recipe information in a list form.
}
\description{
Gives full recipe information with input of ingredients and dietary restriction
}
\examples{
FoodYouHave <- c("chickpeas", "tomato", "spinach")
recipe <- get_info(FoodYouHave, "vegan")

print(recipe)

}
