% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantryPackage.R
\docType{package}
\name{Pantry}
\alias{Pantry}
\title{This package produces recipes based on what food you already have}
\description{
The package contains two functions: 1) get_id() and 2) get_info()
}
\section{Get ID}{

\emph{get_id} produces ID numbers of recipes based on an ingredient list
input
}

\section{Recipe Info}{

\emph{get_info} produces full recipe information based on input of ingredients
 and dietary restrictions
}

