% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infofun.R
\name{get_info}
\alias{get_info}
\title{Gives an output of ingredients needed for recipe and a url to the recipe with
input of ingredients and dietary restriction}
\usage{
get_info(ingredients, diet = "None")
}
\arguments{
\item{ingredients}{A character vector of individual ingredients;
no limit on length}

\item{diet}{A character of dietary restrictions. By default it is NULL.
If diet label is longer than one word, hyphenate it
(e.g., gluten free = Gluten-Free)}
}
\value{
Gives recipe information in a list form.
}
\description{
Gives an output of ingredients needed for recipe and a url to the recipe with
input of ingredients and dietary restriction
}
\examples{
FoodYouHave <- c("chickpeas", "tomato", "spinach")
recipe <- get_info(FoodYouHave, "Vegan")

print(recipe)

}
