% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantryPackage.R
\docType{package}
\name{Pantry}
\alias{Pantry}
\title{This package produces recipes based on what food you already have}
\description{
The package contains one function: get_info()
}
\section{Recipe Info}{

\emph{get_info} produces full recipe information based on input of ingredients
 and dietary restrictions
}

