% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantryPackage.R
\docType{package}
\name{Pantry}
\alias{Pantry}
\title{This package produces recipes based on what food you already have}
\description{
The package contains two functions: get_info() and whatsinmypantry()
}
\section{Recipe Info}{

\emph{get_info} produces full recipe information based on input of
ingredients and dietary restrictions
}

\section{What's in my pantry}{

\emph{whatsinmypantry} launches the "What's in my pantry?" shiny app in a
browser window
}

